check if BLAST already done
add: exists(seq_id, threshold): quickly check the savefile to see if a sequence identifier (say, a header) already exists <threshhold> amount of times. If it does, return the count. Else, return False. Used to skip sequences that are already blasted
Issue: doesn't check which blast-db-matrix results are in the savefile. Could be changed to take a list of parameters to check each line in the savefile for.
Could later be changed to use SQL to check 'our' database for alreadt existing entries